app/requirements.txt
Flask==2.1.1
PyJWT==2.8.0
requests==2.31.0
flask-restplus==0.13.0